Removing AiiDA implicit dependency from qe_tools
The code has been moved into AiiDA with PR
aiidateam/aiida-core#3216
(inside `aiida.tools`).

This was anyway needed also because the imports
used in `qe_tools` were only working from AiiDA 0.12.x or
earlier, but were failing in AiiDA 1.0.

In this way, we make this package independent of AiiDA.
As such, this commit also fixes #3