Add utility to convert cell into parameters.
New functionality:
Adds a 'get_parameters_from_cell' function, which converts the
cell into either the A, B, C, cosAB, cosAC, cosBC or the celldm(*)
parameters. The function relies on 'ibrav' being specified by the
user. The existing 'get_cell_from_parameters' is used to check
the results - this detects cases where the given 'ibrav' and
'cell' do not match.

Naming:
The 'get_parameters_from_cell' is put in a new 'generators' module.
The idea behind this name is that (input-)generators are opposite
to the existing (output-)parsers. It's not a great naming scheme,
so better ideas are very welcome.

Scaffolding:
The 'pytest-cases' library is used to create test cases, from the
existing reference data for the parser.
Mypy is added to the pre-commit hooks for static type checking.