Implement parsing for 'ibrav=-13'.
Because the meaning of ibrav=-13 changed in QE version 6.4.1,
this adds a way of dealing with different QE versions. The
'parse_version' function turns a string (e.g. '6.4.1') into
a 'packaging.version.Version'. By default the latest version is
used, which is implemented as a singleton object that compares
as greater than any 'Version' object.
In the interface, the version can be specified as the optional
keyword-only argument 'qe_version'. Because this option needs to
be available both at the top-level __init__ of the 'QeInputFile'
and on the free functions (e.g. 'get_cell_from_parameters'), it
accepts either a string or an already-parsed version. This allows
immediately parsing the version in 'QeInputFile.__init__', while
also accepting string input in 'get_cell_from_parameters'.
The way this is implemented is by making the 'parse_version'
idempotent, meaning that it will not change an already-parsed
version.

In the tests, the 'qe_version' can be passed as a keyword to
'singletest'. However, these tests currently cannot auto-generate
the reference file.