Fix OLD version of ibrav=-13.
There was an error in the pw.x documentation, according to which
the ibrav=-13 parsing was implemented. As a result, the "OLD"
version of the parsing was incorrect, which is now fixed.
Also, the QE version in which the new ibrav=-13 was introduced
is 6.5, not 6.4.1. The conditional for this switch is also fixed.