Separate public / private interface. (#43)
Separate the public and private interface, as detailed in
issues #41,42,25.
The duplicated code from CpInputFile and PwInputFile has been
moved into the common base, which was renamed and made private.
The constants are now given as a SimpleNamespace, instead of
defined at the top-level of their module. The reason for this is
that we want to be compatible with adding version-specific
constants later on. This could now be done by adding a
'get_constants' function that takes the QE version as input,
and returns the SimpleNamespace of corresponding constants.