✨ Add basic `stdout` parsing functionality
Currently the `PwStdoutParser` nor its parent class `BaseOutputFileParser` parse any
of the Quantum ESPRESSO `stdout`. Here we add the `BaseStdoutParser` class that has
some very minimal parsing features for parsing the code name and version, as well as the
walltime of the calculation. A utility function is also added to convert the Quantum
ESPRESSO walltime output into seconds.

The constructor of the `BaseOutputFileParser` is also adapted to remove the default
to `None`. Since there is currently no way of providing the output content after
construction, it should be provided at that time for the class to be useful.