🔧 Sync package devops to `python-copier` template
Since I want to control the devops of the package I maintain in a single repository,
we sync the devops in this package to the `python-copier` template:

https://github.com/mbercx/python-copier

Using the `copier copy` command. Some notes:

1. This commit only shows the changes introduces by the sync, and is commited with
   `--no-verify` to avoid muddying the actual changes with the changes introduced by
   the switch in formatter/linter (options).

2. The changes will most likely break the pre-commit, tests and existing GitHub actions:

   * The pre-commit will be fixed in the next commit by running the new `hatch` scripts.
   * Tests and GitHub actions are removed and reinstated in a future commit by
     updating the `python-copier` template package and using `copier update` here.

3. We also switch from documentation tool from MyST/Sphinx to MkDocs. We'll see if we
   run into limitations with this much simpler documentation tool.

This is a good test to see if it is possible and relatively painless to sync a package
with a more mature devops configuration to the `python-copier` template package.