🔧 Run new Hatch `precommit` setup
With the new Hatch command, run the `precommit` based on Hatch formatting and linting,
using Ruff under the hood.

Several linting rules are ignored. The motivation will be added to the template package
and then integrated here when we update the template.

Note that we also exlcude the tests for now. The number of fixes was already very large,
we will run linting for the tests again when we reinstate the GitHub Action CI.