🔧 Update package template to v0.9.0
See

https://github.com/mbercx/python-copier/releases/tag/v0.9.0

Main changes include the reintroduction of the test runs to the CI. By doing so, we
caught some bugs that were introduced by updating the formatter/linter to Ruff.

Some of the tests still rely on `unittest`-based approaches, which the linter is not
a fan of. Moreover, the tests also still use e.g. `pytest_cases` that can be replaced by
`@pytest.mark.parametrize`. We add these rules to the Ruff ignore list for the tests,
and stick with the outdated depedencies for now, to avoid having to rewrite a large part
of the tests which we may no longer use in the future.