🔧 Add `.jupytext-sync-ipynb` to `.gitignore`
To write the documentation, it would be ideal to be able to just write in a Jupyter
notebook, so you can just run the cells locally and test the content. However, the
`.ipynb` format is JSON, which is very cumbersome and leads to large and hard to read
diffs.

Jupytext sync is a tool for writing notebooks in `.ipynb` format, and syncing them to a
`.md` file, see:

https://marketplace.visualstudio.com/items?itemName=caenrigen.jupytext-sync

The Markdown format is also used by our supported documentation tool MkDocs. Here we
add the directory that Jupytext sync creates to the `.gitignore` to avoid committing
the `.ipynb` files it generates.