✨ `PwOutput`: add `from_files` method
Currently, the user can specify which file they want to parse from via the `filetype`
input argument of the `from_dir` method. However, this is a bit unintuitive, and
it's not immediately clear what the options are (they are also missing from the
docstring). Morever, it doesn't fit various use cases, such as e.g. parsing input
from a file handle.

Here we remove the `filetype` input and instead add a `from_files` method which
parses the outputs directly from the provided files with the corresponding parser class.
The `from_dir` method is also adapted to use this method.

We decided to add `*` to the function signature to force the user to be explicit in
what files they want to import from with keyword arguments.