👌 `BaseOutput`: rename `outputs` to `raw_outputs`
The current `outputs` input of the `BaseOutput` constructor contains all the "raw"
parsed data, which is a very large dictionary. This typically is not what a user would
want.

In preparation for a more user-friendly approach, we rename the `outputs` class variable
to `raw_outputs`, making the expected contents more clear.

We also adapt the `BaseOutput` constructor to _require_ the `raw_outputs` input
argument; instantiating the class without it does not make much sense at the moment.
Moreover, the `PwOutput` defines its own constructor, but the only thing it does is
call the parent one. Hence we can just remove it.