✨ Converters: add basic design implementation
Add several new classes that implement basic parsing/conversion features for the `pw.x`
outputs. The full motivation is too large to be described in a commit message, we
instead provide a new "Outputs" page in the design documentation that provides a first
draft with the considerations that went into the changes here.

What we will highlight is that the implementation covers several use cases:

1. Extract relevant outputs from the raw `pw.x` output using JSON queries written with
   `glom`.
2. Convert the outputs, both their units, and into instances of the tools we want to
   support: AiiDA, ASE and pymatgen.
3. Allow "default" extractions for simple data (e.g. a float like Fermi energy), that
   can be shared by all converters.
4. Basic conversion simply passes extracted data into the constructor of the output
   class; more complex cases like AiiDA's `StructureData` cannot be directly constructed
   from the constructor. This points to poor design here, but we can still support this
   by allowing the first element in the `(<output_converter>, <glom_spec>)` tuple to
   be a function.

A `get_output` method is implemented on the `PwOutput`, which is the main user-facing
interface for all these features.