👌 `BaseConverter`: remove `numpy` dependency
The goal for the `BaseConverter` is to provide basic Python types for each of the
outputs. However, the current `structure` output also contains `numpy` arrays.

Here we avoid using `numpy` in the `BaseConverter` class. This has the added benefit
that it will be easier in case we want to make `numpy` an optional dependency.