❌ Scope: drop support for QE versions <7.0
In line with the support statement added to the "Design - Scope" section, we drop
support for any Quantum ESPRESSO version below 7.0.

We considered keeping XML parsing support. But it already became clear when just
trying to parse the Fermi energy that older versions of the XML are still quite limited.
Instead of providing "partial" support, we remove support entirely and make it clear
what versions _are_ supported. In fact, support may be reduced in scope further if we
find other cases that are hard to work around.