🐛 Converters: fix atomic species spec
The current `glom` spec for extracting the "symbols" is incorrect. The
`atomic_species` output in the XML only has the list of "kinds" in the structure. What
we are looking for here is the "name" of each atomic position instead.

We add a `json_serializer` fixture to the tests, specifically for the ASE case where the
`Atoms.todict` method does not produce a JSON-serializable result (it contains numpy
arrays).