🧪 `json_serializer`: extend functionality and reach
Due to small differences in float values and how integers are serialized, the CI tests
are failing when the data regression is compared with the YAML file. Here we first adapt
the `json_serializer` fixture to also support serializing floats/integers into more
robust values (always float, rounded to 5 decimal numbers).

In order to have to add and apply the `json_serializer` everywhere it is needed, we
simply create a small `robust_data_regression_check` fixture that we can use for most if
not all of our regressions checks.