✨ `BaseOutput`: add `get_output_from_spec` method
A common use case is that we might not support parsing for a certain output (yet). Or
users want to get the output in a different structure/format. In principle, they can
already do this quite easily using `glom` and the `raw_output`. But to make this option
more accessible and findable, we add a `get_output_from_spec` method that exactly uses
the provided `spec` input to extract data from the `raw_output` with `glom`.