✨ Parsers: make classes stateless
In the current implementation, we need to

1. instantiate a parser class.
2. run the `.parse()` method on the `string` content.
3. extract the raw outputs from the `dict_out` attribute.

However, there is no need to store any data on a parser class. Instead, we can convert
the parser classes into stateless objects, which have to implement a single `parse`
method with the following signature:

    def parse(content: str):

        parsed_data: dict = ...
        return parsed_data

This make their usage and implementation more transparent: no more `dict_out` attribute
that suddenly appears in the `parse` method of a child class. Moreover, it means we can
obtain the parsed data in a single step, e.g.:

    parsed_data = PwXMLParser.parse(content)

One motivation for having a stateful class might be performance: by storing the raw
content on the class and only parsing it later, we can selectively parse what is needed.
However, parsing the XML is most likely the most expensive operation we'll have to do
here, in part because of the validation done by `xmlschema`. We're not sure if partially
parsing the XML is an option, or we would lose the validation. There is also no
performance bottleneck at this stage, even for larger XML files the parsing only takes
50 ms.

Since the parser classes are not part of the public API at this point, we can still
come back on this point at a later date. Moreover, we can still keep the static methods
in place, i.e. make changes to support partial parsing in a backwards-compatible manner.

Finally, we rename the `string` input to `content`, which is of type `str`.