👌 `PwXMLParser`: improve error message for XML parsing failure
In case the parsing of the XML fails, the user is now confronted with a rather large
and difficult to understand traceback. Here we catch the `ElementTree.ParseError` and
return a simple `ValueError` with some details on what to check.

For now, a `ValueError` should be sufficient, even though it is perhaps not descriptive
enough. We could decide to make our own custom error messages, but this practise should
be considered more broadly.

In order to keep the traceback minimal, we add `from None` after the `raise` statement.
This will remove information on _where_ the parsing went wrong, but frankly if the XML
is incorrect the user should try to investigate and fix the issue manually.