🔄 Outputs: separate extraction from conversion
Working on 1e3b38cf3a4f58c9af529f70bc4418961c6372bb highlighted a flaw in our current
design. The "base outputs" of the `PwOutput` were added to the `BaseConverter` class.
This has a few issues:

1. Obviously, when extending to other codes we would need to also have corresponding
   `BaseConverter` classes (e.g. `BandsBaseConverter`). Additionally, we'd need one
   converter class for each code for each supported tool. That's a lot of classes.
2. The logic regarding extracting base outputs for `pw.x` is separated. We have to
   make sure that the glom spec defined on `BaseConverter` match with what is parsed
   into the raw outputs on `PwOutput`.
3. It muddles the line between an "extractor" and a "converter". What we were calling
   a converter was actually an extractor + converter. We have considered calling the
   classes "extractors" for this very reason.

Here, we move the `BaseConverter.output_mapping` to `PwOutput.output_spec_mapping`.
We also remove the output classes: since the base outputs are meant to be only basic
Python types we don't need them so far (and shouldn't in the future).

The `BaseConverter.get_output` method is renamed to `convert`, which is closer to what
it now does via its subclasses: convert the base output into an output of the
corresponding library. The `output_mapping` for the subclasses for each library is
now called `conversion_mapping`. Instead of writing a `glom` spec that extracts data
from the raw output, it defined a much simpler one that converts the base output into
the inputs for the library one, or the conversion function in the more complicated case
of AiiDA's `StructureData`.

This has some advantages:

- We no longer need to make a separate converter class for each QE code. All the logic
  for the various outputs can sensibly be added to a single class. I.e. [1] above is
  solved. So are [2] and [3].
- The `list_outputs` method is now simplified, and more sensible: it's not extracting
  the list of outputs from another "converter" class in another module.
- Less code repetition: the "extraction" logic is no longer necessary in each of the
  converter classes.

However, we also introduce some disadvantages:

1. When writing an element for the `conversion_mapping`, you need to know the structure
   of the data in the base output.
2. We add another step to the process of getting to the converted outputs.
3. For the "converted" outputs, the logic of how they are obtained is separated.

[1] can potentially be solved by good schema/model definitions of the supported outputs
for conversion. [2] and [3] are a reasonable sacrifice, because the whole point is to
separate these two concepts.