✨ `BaseOutput`: add `outputs` namespace
Add an `outputs` namespace on the `BaseOutput` class that allows users to use
tab-completion to obtain the outputs that can be successfully retrieved.

The `outputs` namespace is implemented as a cached property, since the `BaseOutput`
classes should not allow for state changes after construction: they only have `from_X`
class constructor methods and getter-type methods for retrieving data. This should
remain the case.

For now, we simply skip outputs that cannot be retrieved. The idea is that this
approach for obtaining the outputs is meant for interactive usage. We're still
considering where and how to raise a warning in case outputs are missing.

The `list_outputs` method is also moved closer to this one, since they have the same
scope.