✨ `PwOutput`: add `forces` and `stress` outputs
Note that the XML units are:

- forces: Ha / Bohr
- stress: Ha / Bohr^3

Hence we need to be careful when converting to our (currently) desired units:

- forces: eV / Angstrom
- stress: GPa

For now, we also avoid using `numpy` in the base outputs. We should evaluate soon
whether or not avoiding this dependency is feasible/desirable, because it does
make the code less readable.