✨ `DosOutput`: add first version
Add the first version of the `DosOutput` class, a parser for `dos.x` outputs, via a new
file parser for the `.dos` output file, supporting the various spin cases.

As a demonstrator, the `DosOutput` class also tries to find the XML file, and extracts
the spin type from it. A typical Quantum ESPRESSO user will run `pw.x` then `dos.x` in
the same directory, so both files _should_ be present. This ability for the various
output classes to use the file parsers of other codes and extract only the information
they need with `glom` queries is very flexible, and demonstrates one of the reasons
behind the design.

Using `glom` queries, we can also offer the user the outputs in two ways: having the
energy, DOS, integrated DOS etc at the top-level, immediately accessible via
tab-completion in the `outputs` namespace, or gathered in the `full_dos` output. This
doesn't actually duplicate any data: the "raw" outputs are simply queried differently
for the desired "base" outputs. The original motivation for adding the `full_dos` is
to then allow conversion into the types of our supported libraries, e.g. an `ArrayData`
for AiiDA.

Note that we _do_ use numpy here, despite a desire to avoid dependency on this package.
Frankly, the code without using number becomes more complicated and less readable. We
have to see if it is worth avoiding numpy, since it is also ubiquitous in our ecosystem.