✨ `AiiDAConverter`: add `full_dos` conversion
As a demonstrator of the motivation of having a `full_dos` output, we add a conversion
mapping from `full_dos` to an AiiDA `XyData`. It should be clear that this mapping can
be improved: we're missing a true `DosData` data type, which could also store the Fermi
energy. However, here we show the advantage of separating the "extraction" to base
outputs and "conversion" into library ones: we can have a single converter class that
takes care of the conversion of outputs of the various Quantum ESPRESSO codes.

While working on this, some other considerations regarding the design came to mind.
These are added to corresponding design section.