👌 `BaseOutput`: add `only_available` argument to `list_outputs`
Currently, the `BaseOutput.list_outputs` method list all of the supported outputs,
whether they are available or not in the raw outputs. This has several downsides:

- It can be confusing for the user: if they first list the output, and then try to
  retrieve a supported but non-available one via `get_output` method.
- How can I programmatically see what outputs are actually available? Currently only
  the `outputs` namespace offers that feature, but you'd have to inspect the
  attributes or something similar, which is more convoluted.

Here, we adapt the `list_outputs` method to only show the available outputs by default.
To still support showing all supported outputs, we add the `only_available` input
argument that is `True` by default. This new signature is then also used internally for
populating the `outputs` namespace.