✨ `BaseOutput`: add the `get_output_dict` method
In many cases of programmatic usage, you want to obtain all or a subset of outputs in
one step, without having to list the outputs first and retrieve them one by one.

Here we add the `BaseOutput.get_output_dict` method for this use case, which by default
returns a dictionary of all available outputs. A subset can be selected via the `names`
argument, and the outputs can be converted to the type of a desired library using `to`.